Extract duplicated test helpers into a shared scope
